refactor: move v1-specific code to v1 package
