+++ annotations/crd-annotations/src/main/java/io/dekorate/crd/v1/decorator/AddSubresourcesDecorator.java
@@ -15,29 +15,26 @@
  * 
 **/
 
-package io.dekorate.crd.decorator;
+package io.dekorate.crd.v1.decorator;
 
 import io.dekorate.kubernetes.decorator.Decorator;
 import io.fabric8.kubernetes.api.model.apiextensions.v1.CustomResourceDefinitionVersionFluent;
-import io.fabric8.kubernetes.api.model.apiextensions.v1.JSONSchemaProps;
 
-public class AddSchemaToCustomResourceDefinitionVersionDecorator
-    extends CustomResourceDefinitionVersionDecorator<CustomResourceDefinitionVersionFluent<?>> {
+public class AddSubresourcesDecorator extends CustomResourceDefinitionVersionDecorator<CustomResourceDefinitionVersionFluent<?>> {
 
-  private JSONSchemaProps schema;
+	public AddSubresourcesDecorator(String name, String version) {
+		super(name, version);
+	}
 
-  public AddSchemaToCustomResourceDefinitionVersionDecorator(String name, String version, JSONSchemaProps schema) {
-    super(name, version);
-    this.schema = schema;
-  }
+	@Override
+	public void andThenVisit(CustomResourceDefinitionVersionFluent<?> spec) {
+    if (!spec.hasSubresources()) {
+      spec.withNewSubresources().endSubresources();
+    }
+	}
 
-  @Override
-  public void andThenVisit(CustomResourceDefinitionVersionFluent<?> version) {
-    version.withNewSchema().withOpenAPIV3Schema(schema).endSchema();
-  }
-
-  @Override
-  public Class<? extends Decorator>[] after() {
+	@Override
+	public Class<? extends Decorator>[] after() {
     return new Class[]{AddCustomResourceDefinitionVersionDecorator.class};
-  }
+	}
 }
