+++ annotations/crd-annotations/src/main/java/io/dekorate/crd/v1/decorator/SetServedVersionDecorator.java
@@ -15,18 +15,21 @@
  * 
 **/
 
-package io.dekorate.crd.decorator;
+package io.dekorate.crd.v1.decorator;
 
-import io.dekorate.kubernetes.decorator.NamedResourceDecorator;
-import io.fabric8.kubernetes.api.model.ObjectMeta;
+import io.fabric8.kubernetes.api.model.apiextensions.v1.CustomResourceDefinitionVersionFluent;
 
-public class CustomResourceDefinitionDecorator<T> extends NamedResourceDecorator<T> {
+public class SetServedVersionDecorator extends CustomResourceDefinitionVersionDecorator<CustomResourceDefinitionVersionFluent<?>> {
 
-	public CustomResourceDefinitionDecorator(String name) {
-		super("CustomResourceDefinition", name);
+  private final boolean served;
+
+	public SetServedVersionDecorator(String name, String version, boolean served) {
+		super(name, version);
+		this.served = served;
 	}
 
 	@Override
-	public void andThenVisit(T item, ObjectMeta resourceMeta) {
+	public void andThenVisit(CustomResourceDefinitionVersionFluent<?> version) {
+    version.withServed(served);
 	}
 }
